feat(profile): selector avatar IA — colibrí o planta de maíz (#307)
Operator decisión 2026-05-27: dar opción al usuario en Personalización
para cambiar el colibrí (default) por una planta de maíz si prefiere.

Arquitectura drop-in:
- src/components/ChagraAgentAvatar.jsx → wrapper que delega según pref
- src/components/ChagraAgentAvatarColibri.jsx → SVG colibrí (era el actual)
- src/components/ChagraAgentAvatarMaiz.jsx → SVG planta de maíz nuevo
- src/hooks/useAgentAvatarType.js → hook con localStorage + event sync
  cross-tab via CustomEvent 'chagra:agent-avatar-changed'

SVG planta de maíz:
- viewBox 200x200 (mismo que colibrí, drop-in compatible)
- Tallo central + nudos + 4 hojas + mazorca lateral con barbas doradas
  + panocha apical con polen
- 4 estados como el colibrí:
  - idle: brisa suave en hojas (2.5s)
  - thinking: barbas de la mazorca ondulan (0.8s) + hojas inquietas
  - speaking: hojas oscilan rápido + panocha vibra
  - listening: planta entera se inclina lateralmente (3s)
- Glow amber cuando glow=true (igual al colibrí)

UI en ProfileScreen → Personalización (al lado de ThemeSelector):
- 2 botones grandes con preview del avatar real en estado 'thinking'
- aria-pressed para a11y
- Cambio inmediato visible en toda la app gracias al CustomEvent

Tests: 4/4 pass (renderiza 2, default colibri, click maiz persiste +
visual, localStorage preselect).
