+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "chagra",
   "private": true,
-  "version": "0.13.21",
+  "version": "0.13.22",
   "type": "module",
   "scripts": {
     "dev": "vite",
